updated web app flask scripts as well as testing shell script
