--- distributed_systems/replicated_load-balanced_services/varnish_caching/web_app/requirements.txt
@@ -1,1 +1,0 @@
-Flask
