+++ distributed_systems/replicated_load-balanced_services/varnish_caching/web_app_1/requirements.txt
@@ -1,0 +1,1 @@
+flask
