+++ distributed_systems/replicated_load-balanced_services/varnish_caching/web_app_2/Dockerfile
@@ -1,0 +1,19 @@
+# use python offical image from dockerhub as base image for this container
+FROM python 
+
+# copy contents of this directory to the /app directory within the container
+COPY . /app
+
+# set working directory of container
+WORKDIR /app
+
+# install the python prerequisites using the requirements.txt file
+RUN pip install -r requirements.txt
+
+# set the port the container will listen for requests on to 5001
+EXPOSE 5001
+
+# use python to execute the flask web application
+ENTRYPOINT [ "python" ]
+CMD [ "web_app_2.py" ]
+
